See description:
Add writeByte for both I2C and SPI - needed by the ADS1219 on I2C
Add support for I2C restart vs. stop (many devices require a restart between set-register-address and read-register-value). This is I2C-specific and may be needed on non-Arduino platforms too. Added setStop and getStop.
Corrected readRegisterRegion: the beginTransmission and endTransmission should be inside the if(bFirstInter).